refactor: extract api documentation into interfaces
