+++ src/main/java/com/luminabackend/controllers/account/AccountController.java
@@ -1,15 +1,9 @@
-package com.luminabackend.controllers;
+package com.luminabackend.controllers.account;
 
 import com.luminabackend.models.user.User;
 import com.luminabackend.models.user.dto.user.UserLoginDTO;
 import com.luminabackend.services.TokenService;
-import com.luminabackend.utils.errors.GeneralErrorResponseDTO;
 import com.luminabackend.utils.security.TokenDTO;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -22,28 +16,14 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping
-public class AccountController {
+public class AccountController implements AccountControllerDocumentation {
     @Autowired
     private AuthenticationManager manager;
 
     @Autowired
     private TokenService tokenService;
 
-    @Operation(summary = "Login with email and password")
-    @ApiResponses(value = {
-            @ApiResponse(
-                    responseCode = "200",
-                    description = "Successfully login",
-                    content = { @Content(
-                            mediaType = "application/json",
-                            schema = @Schema(implementation = TokenDTO.class)) }),
-            @ApiResponse(
-                    responseCode = "401",
-                    description = "Unauthorized. Incorrect or invalid credentials",
-                    content = { @Content(
-                            mediaType = "application/json",
-                            schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
-    })
+    @Override
     @PostMapping("/login")
     public ResponseEntity<TokenDTO> login(@Valid @RequestBody UserLoginDTO data){
         String email = data.email().trim();
