+++ src/main/java/com/luminabackend/controllers/classroom/ClassroomControllerDocumentation.java
@@ -1,20 +1,17 @@
-package com.luminabackend.controllers;
+package com.luminabackend.controllers.classroom;
 
-import com.luminabackend.models.education.classroom.*;
-import com.luminabackend.models.user.dto.user.UserAccessDTO;
-import com.luminabackend.services.ClassroomService;
-import com.luminabackend.services.AccessService;
-import com.luminabackend.services.TokenService;
+import com.luminabackend.models.education.classroom.ClassroomGetDTO;
+import com.luminabackend.models.education.classroom.ClassroomPostDTO;
+import com.luminabackend.models.education.classroom.ClassroomPutDTO;
+import com.luminabackend.models.education.classroom.ClassroomWithRelationsDTO;
 import com.luminabackend.utils.errors.GeneralErrorResponseDTO;
 import com.luminabackend.utils.errors.ValidationErrorResponseDTO;
-import com.luminabackend.utils.security.PayloadDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.validation.Valid;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
@@ -24,9 +21,6 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import java.util.UUID;
 
-import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
-import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
-
 @ApiResponses(value = {
         @ApiResponse(
                 responseCode = "401",
@@ -35,18 +29,7 @@ import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
                         mediaType = "application/json",
                         schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
 })
-@RestController
-@RequestMapping("/classroom")
-public class ClassroomController {
-    @Autowired
-    private ClassroomService classroomService;
-
-    @Autowired
-    private TokenService tokenService;
-
-    @Autowired
-    private AccessService accessService;
-
+public interface ClassroomControllerDocumentation {
     @Operation(summary = "Get a list of all classrooms based on user access")
     @ApiResponses(value = {
             @ApiResponse(
@@ -55,13 +38,7 @@ public class ClassroomController {
                     content = { @Content(mediaType = "application/json",
                             schema = @Schema(implementation = ClassroomGetDTO.class)) })
     })
-    @PreAuthorize("hasRole('ADMIN') or hasRole('PROFESSOR') or hasRole('STUDENT')")
-    @GetMapping("/all")
-    public ResponseEntity<List<ClassroomGetDTO>> getAllClassrooms(@RequestHeader("Authorization") String authorizationHeader) {
-        PayloadDTO payloadDTO = tokenService.getPayloadFromAuthorizationHeader(authorizationHeader);
-        List<Classroom> classrooms = classroomService.getClassroomsBasedOnUserAccess(new UserAccessDTO(payloadDTO));
-        return ResponseEntity.ok(classrooms.stream().map(ClassroomGetDTO::new).toList());
-    }
+    ResponseEntity<List<ClassroomGetDTO>> getAllClassrooms(@RequestHeader("Authorization") String authorizationHeader);
 
     @Operation(summary = "Get a paginated list of classrooms based on user access")
     @ApiResponses(value = {
@@ -71,13 +48,7 @@ public class ClassroomController {
                     content = { @Content(mediaType = "application/json",
                             schema = @Schema(implementation = ClassroomGetDTO.class)) })
     })
-    @PreAuthorize("hasRole('ADMIN') or hasRole('PROFESSOR') or hasRole('STUDENT')")
-    @GetMapping
-    public ResponseEntity<Page<ClassroomGetDTO>> getPaginatedClassrooms(Pageable page, @RequestHeader("Authorization") String authorizationHeader) {
-        PayloadDTO payloadDTO = tokenService.getPayloadFromAuthorizationHeader(authorizationHeader);
-        Page<Classroom> classrooms = classroomService.getPaginatedClassroomsBasedOnUserAccess(new UserAccessDTO(payloadDTO), page);
-        return ResponseEntity.ok(classrooms.map(ClassroomGetDTO::new));
-    }
+    ResponseEntity<Page<ClassroomGetDTO>> getPaginatedClassrooms(Pageable page, @RequestHeader("Authorization") String authorizationHeader);
 
     @Operation(summary = "Get a classroom by its id based on user access")
     @ApiResponses(value = {
@@ -106,15 +77,9 @@ public class ClassroomController {
                             mediaType = "application/json",
                             schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
     })
-    @PreAuthorize("hasRole('ADMIN') or hasRole('PROFESSOR') or hasRole('STUDENT')")
-    @GetMapping("/{classroomId}")
-    public ResponseEntity<ClassroomGetDTO> getClassroom(
+    ResponseEntity<ClassroomGetDTO> getClassroom(
             @PathVariable UUID classroomId,
-            @RequestHeader("Authorization") String authorizationHeader) {
-        PayloadDTO payloadDTO = tokenService.getPayloadFromAuthorizationHeader(authorizationHeader);
-        Classroom classroom = classroomService.getClassroomBasedOnUserAccess(classroomId, new UserAccessDTO(payloadDTO));
-        return ResponseEntity.ok(new ClassroomGetDTO(classroom));
-    }
+            @RequestHeader("Authorization") String authorizationHeader);
 
     @Operation(summary = "Get the classroom professor and a list of its students")
     @ApiResponses(value = {
@@ -143,17 +108,9 @@ public class ClassroomController {
                             mediaType = "application/json",
                             schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
     })
-    @PreAuthorize("hasRole('ADMIN') or hasRole('PROFESSOR') or hasRole('STUDENT')")
-    @GetMapping("/{classroomId}/members")
-    public ResponseEntity<ClassroomWithRelationsDTO> getClassroomWithRelations(
-        @PathVariable UUID classroomId,
-        @RequestHeader("Authorization") String authorizationHeader) {
-        PayloadDTO payloadDTO = tokenService.getPayloadFromAuthorizationHeader(authorizationHeader);
-        Classroom classroom = classroomService.getClassroomById(classroomId);
-        accessService.checkAccessToClassroom(classroom, new UserAccessDTO(payloadDTO));
-        ClassroomWithRelationsDTO fullClassroom = classroomService.getClassroomWithRelations(classroom);
-        return ResponseEntity.ok(fullClassroom);
-    }
+    ResponseEntity<ClassroomWithRelationsDTO> getClassroomWithRelations(
+            @PathVariable UUID classroomId,
+            @RequestHeader("Authorization") String authorizationHeader);
 
     @Operation(summary = "Create a new classroom")
     @ApiResponses(value = {
@@ -169,18 +126,9 @@ public class ClassroomController {
                             mediaType = "application/json",
                             schema = @Schema(implementation = ValidationErrorResponseDTO.class)) }),
     })
-    @PreAuthorize("hasRole('ADMIN')")
-    @PostMapping
-    public ResponseEntity<ClassroomGetDTO> saveClassroom(
+    ResponseEntity<ClassroomGetDTO> saveClassroom(
             @Valid @RequestBody ClassroomPostDTO classroomPostDTO,
-            @RequestHeader("Authorization") String authorizationHeader) {
-        Classroom savedClassroom = classroomService.save(classroomPostDTO);
-        return ResponseEntity
-                .created(linkTo(methodOn(ClassroomController.class)
-                        .getClassroom(savedClassroom.getId(), authorizationHeader))
-                        .toUri())
-                .body(new ClassroomGetDTO(savedClassroom));
-    }
+            @RequestHeader("Authorization") String authorizationHeader);
 
     @Operation(summary = "Edit a classroom by its id")
     @ApiResponses(value = {
@@ -208,14 +156,9 @@ public class ClassroomController {
                             mediaType = "application/json",
                             schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
     })
-    @PreAuthorize("hasRole('ADMIN')")
-    @PutMapping("/{classroomId}")
-    public ResponseEntity<ClassroomGetDTO> editClassroom(
+    ResponseEntity<ClassroomGetDTO> editClassroom(
             @PathVariable UUID classroomId,
-            @Valid @RequestBody ClassroomPutDTO classroomPutDTO) {
-        Classroom classroom = classroomService.edit(classroomId, classroomPutDTO);
-        return ResponseEntity.ok(new ClassroomGetDTO(classroom));
-    }
+            @Valid @RequestBody ClassroomPutDTO classroomPutDTO);
 
     @Operation(summary = "Delete a classroom and its dependencies by classroom id")
     @ApiResponses(value = {
@@ -235,12 +178,7 @@ public class ClassroomController {
                             mediaType = "application/json",
                             schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
     })
-    @PreAuthorize("hasRole('ADMIN')")
-    @DeleteMapping("/{classroomId}")
-    public ResponseEntity<Void> deleteClassroom(@PathVariable UUID classroomId) {
-        classroomService.deleteById(classroomId);
-        return ResponseEntity.noContent().build();
-    }
+    ResponseEntity<Void> deleteClassroom(@PathVariable UUID classroomId);
 
     @Operation(summary = "Add a student to a classroom")
     @ApiResponses(value = {
@@ -280,17 +218,10 @@ public class ClassroomController {
                             mediaType = "application/json",
                             schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
     })
-    @PreAuthorize("hasRole('ADMIN') or hasRole('PROFESSOR')")
-    @PostMapping("/{classroomId}/student/{studentId}")
-    public ResponseEntity<Void> addStudent(
+    ResponseEntity<Void> addStudent(
             @PathVariable UUID classroomId,
             @PathVariable UUID studentId,
-            @RequestHeader("Authorization") String authorizationHeader) {
-        PayloadDTO payloadDTO = tokenService.getPayloadFromAuthorizationHeader(authorizationHeader);
-        Classroom classroom = classroomService.getClassroomBasedOnUserAccess(classroomId, new UserAccessDTO(payloadDTO));
-        classroomService.addStudentToClassroom(studentId, classroom);
-        return ResponseEntity.ok().build();
-    }
+            @RequestHeader("Authorization") String authorizationHeader);
 
     @Operation(summary = "Remove a student from classroom")
     @ApiResponses(value = {
@@ -327,10 +258,5 @@ public class ClassroomController {
     public ResponseEntity<Void> removeStudent(
             @PathVariable UUID classroomId,
             @PathVariable UUID studentId,
-            @RequestHeader("Authorization") String authorizationHeader) {
-        PayloadDTO payloadDTO = tokenService.getPayloadFromAuthorizationHeader(authorizationHeader);
-        Classroom classroom = classroomService.getClassroomBasedOnUserAccess(classroomId, new UserAccessDTO(payloadDTO));
-        classroomService.removeStudentFromClassroom(studentId, classroom);
-        return ResponseEntity.noContent().build();
-    }
+            @RequestHeader("Authorization") String authorizationHeader);
 }
