+++ src/main/java/com/luminabackend/controllers/material/MaterialControllerDocumentation.java
@@ -1,38 +1,27 @@
-package com.luminabackend.controllers;
+package com.luminabackend.controllers.material;
 
-import com.luminabackend.exceptions.MissingFileException;
 import com.luminabackend.models.education.material.Material;
 import com.luminabackend.models.education.material.MaterialGetDTO;
 import com.luminabackend.models.education.material.MaterialPostDTO;
 import com.luminabackend.models.education.submission.SubmissionGetDTO;
-import com.luminabackend.models.user.dto.user.UserAccessDTO;
-import com.luminabackend.services.*;
 import com.luminabackend.utils.errors.GeneralErrorResponseDTO;
 import com.luminabackend.utils.errors.ValidationErrorResponseDTO;
-import com.luminabackend.utils.security.PayloadDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.ByteArrayResource;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.mongodb.gridfs.GridFsResource;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.util.List;
 import java.util.UUID;
-
-import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
-import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 
 @ApiResponses(value = {
         @ApiResponse(
@@ -54,21 +43,7 @@ import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
                         mediaType = "application/json",
                         schema = @Schema(implementation = GeneralErrorResponseDTO.class))})
 })
-@RestController
-@RequestMapping("/classroom/{classroomId}/material")
-public class MaterialController {
-    @Autowired
-    private MaterialService materialService;
-
-    @Autowired
-    private FileStorageService fileStorageService;
-
-    @Autowired
-    private TokenService tokenService;
-
-    @Autowired
-    private AccessService accessService;
-
+public interface MaterialControllerDocumentation {
     @Operation(summary = "Get a list of materials from a classroom")
     @ApiResponses(value = {
             @ApiResponse(
@@ -83,14 +58,9 @@ public class MaterialController {
                             mediaType = "application/json",
                             schema = @Schema(implementation = GeneralErrorResponseDTO.class))}),
     })
-    @PreAuthorize("(hasRole('ADMIN') or hasRole('PROFESSOR') or hasRole('STUDENT')) and @resourceAccess.verifyClassroomAccess(#authorizationHeader, #classroomId)")
-    @GetMapping("/all")
-    public ResponseEntity<List<Material>> getAllClassroomMaterials(
+    ResponseEntity<List<Material>> getAllClassroomMaterials(
             @PathVariable UUID classroomId,
-            @RequestHeader("Authorization") String authorizationHeader) {
-        List<Material> materials = materialService.getAllMaterials(classroomId);
-        return ResponseEntity.ok(materials);
-    }
+            @RequestHeader("Authorization") String authorizationHeader);
 
     @Operation(summary = "Get a paginated list of materials from a classroom")
     @ApiResponses(value = {
@@ -106,15 +76,10 @@ public class MaterialController {
                             mediaType = "application/json",
                             schema = @Schema(implementation = GeneralErrorResponseDTO.class))}),
     })
-    @PreAuthorize("(hasRole('ADMIN') or hasRole('PROFESSOR') or hasRole('STUDENT')) and @resourceAccess.verifyClassroomAccess(#authorizationHeader, #classroomId)")
-    @GetMapping
-    public ResponseEntity<Page<Material>> getPaginatedClassroomMaterials(
+    ResponseEntity<Page<Material>> getPaginatedClassroomMaterials(
             @PathVariable UUID classroomId,
             Pageable page,
-            @RequestHeader("Authorization") String authorizationHeader) {
-        Page<Material> materials = materialService.getPaginatedClassroomMaterials(classroomId, page);
-        return ResponseEntity.ok(materials);
-    }
+            @RequestHeader("Authorization") String authorizationHeader);
 
     @Operation(summary = "Create a new classroom material")
     @ApiResponses(value = {
@@ -142,30 +107,12 @@ public class MaterialController {
                             mediaType = "application/json",
                             schema = @Schema(implementation = GeneralErrorResponseDTO.class))}),
     })
-    @PreAuthorize("hasRole('PROFESSOR') and @resourceAccess.verifyClassroomAccess(#authorizationHeader, #classroomId)")
-    @PostMapping(consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public ResponseEntity<MaterialGetDTO> createMaterial(
+    ResponseEntity<MaterialGetDTO> createMaterial(
             @PathVariable UUID classroomId,
             @RequestPart("material") MaterialPostDTO materialPost,
             @RequestPart("file") MultipartFile file,
-            @RequestHeader("Authorization") String authorizationHeader) throws IOException {
-        if (file == null || file.isEmpty()) {
-            throw new MissingFileException("Missing material file");
-        }
-
-        PayloadDTO payloadDTO = tokenService.getPayloadFromAuthorizationHeader(authorizationHeader);
-        Material savedMaterial = materialService.saveMaterial(
-                classroomId,
-                payloadDTO.id(),
-                materialPost.title(),
-                materialPost.description(),
-                file);
-        return ResponseEntity
-                .created(linkTo(methodOn(MaterialController.class)
-                        .getAllClassroomMaterials(classroomId, authorizationHeader))
-                        .toUri())
-                .body(new MaterialGetDTO(savedMaterial));
-    }
+            @RequestHeader("Authorization") String authorizationHeader
+    ) throws IOException;
 
     @Operation(summary = "Delete a classroom material")
     @ApiResponses(value = {
@@ -185,15 +132,10 @@ public class MaterialController {
                             mediaType = "application/json",
                             schema = @Schema(implementation = GeneralErrorResponseDTO.class))}),
     })
-    @PreAuthorize("hasRole('PROFESSOR') and @resourceAccess.verifyClassroomAccess(#authorizationHeader, #classroomId)")
-    @DeleteMapping("/{materialId}")
-    public ResponseEntity<Void> deleteMaterial(
+    ResponseEntity<Void> deleteMaterial(
             @PathVariable UUID classroomId,
             @PathVariable UUID materialId,
-            @RequestHeader("Authorization") String authorizationHeader) {
-        materialService.deleteById(materialId);
-        return ResponseEntity.noContent().build();
-    }
+            @RequestHeader("Authorization") String authorizationHeader);
 
     @Operation(summary = "Get a classroom material file")
     @ApiResponses(value = {
@@ -221,22 +163,12 @@ public class MaterialController {
                             mediaType = "application/json",
                             schema = @Schema(implementation = GeneralErrorResponseDTO.class))}),
     })
-    @PreAuthorize("(hasRole('ADMIN') or hasRole('PROFESSOR') or hasRole('STUDENT')) and @resourceAccess.verifyClassroomAccess(#authorizationHeader, #classroomId)")
-    @GetMapping("/{materialId}/file/{fileId}")
-    public ResponseEntity<ByteArrayResource> downloadMaterialFile(
+    ResponseEntity<ByteArrayResource> downloadMaterialFile(
             @PathVariable UUID classroomId,
             @PathVariable UUID materialId,
             @PathVariable String fileId,
-            @RequestHeader("Authorization") String authorizationHeader) throws IOException {
-        PayloadDTO payloadDTO = tokenService.getPayloadFromAuthorizationHeader(authorizationHeader);
-        materialService.checkAccessToMaterial(materialId, new UserAccessDTO(payloadDTO));
-
-        GridFsResource file = fileStorageService.getFile(fileId);
-        return ResponseEntity.ok()
-                .contentType(MediaType.parseMediaType(file.getContentType()))
-                .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + file.getFilename() + "\"")
-                .body(new ByteArrayResource(file.getInputStream().readAllBytes()));
-    }
+            @RequestHeader("Authorization") String authorizationHeader
+    ) throws IOException;
 
     @Operation(summary = "Get all classroom materials in a zip file")
     @ApiResponses(value = {
@@ -252,16 +184,8 @@ public class MaterialController {
                             mediaType = "application/json",
                             schema = @Schema(implementation = GeneralErrorResponseDTO.class))})
     })
-    @PreAuthorize("(hasRole('ADMIN') or hasRole('PROFESSOR') or hasRole('STUDENT')) and @resourceAccess.verifyClassroomAccess(#authorizationHeader, #classroomId)")
-    @GetMapping("/download")
-    public ResponseEntity<ByteArrayResource> downloadAllMaterialsFromClassroom(
+    ResponseEntity<ByteArrayResource> downloadAllMaterialsFromClassroom(
             @PathVariable UUID classroomId,
             @RequestHeader("Authorization") String authorizationHeader
-    ) throws IOException {
-        ByteArrayResource resource = materialService.getAllMaterialsAsZip(classroomId);
-        return ResponseEntity.ok()
-                .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=classroom-" + classroomId + "-materials.zip")
-                .contentType(MediaType.APPLICATION_OCTET_STREAM)
-                .body(resource);
-    }
+    ) throws IOException;
 }
