+++ src/main/java/com/luminabackend/controllers/professor/ProfessorControllerDocumentation.java
@@ -1,11 +1,8 @@
-package com.luminabackend.controllers;
+package com.luminabackend.controllers.professor;
 
-import com.luminabackend.exceptions.EntityNotFoundException;
-import com.luminabackend.models.user.Professor;
 import com.luminabackend.models.user.dto.professor.ProfessorGetDTO;
 import com.luminabackend.models.user.dto.user.UserPutDTO;
 import com.luminabackend.models.user.dto.user.UserSignupDTO;
-import com.luminabackend.services.ProfessorService;
 import com.luminabackend.utils.errors.GeneralErrorResponseDTO;
 import com.luminabackend.utils.errors.ValidationErrorResponseDTO;
 import io.swagger.v3.oas.annotations.Operation;
@@ -14,19 +11,13 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.validation.Valid;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
-
-import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
-import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 
 @ApiResponses(value = {
         @ApiResponse(
@@ -36,141 +27,109 @@ import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
                         mediaType = "application/json",
                         schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
 })
-@RestController
-@PreAuthorize("hasRole('ADMIN')")
-@RequestMapping("/professor")
-public class ProfessorController {
-    @Autowired
-    private ProfessorService service;
-
+public interface ProfessorControllerDocumentation {
     @Operation(summary = "Get all professors")
     @ApiResponses(value = {
             @ApiResponse(
                     responseCode = "200",
                     description = "Returns a list with all professors",
-                    content = { @Content(mediaType = "application/json",
-                            schema = @Schema(implementation = ProfessorGetDTO.class)) })
+                    content = {@Content(mediaType = "application/json",
+                            schema = @Schema(implementation = ProfessorGetDTO.class))})
     })
-    @GetMapping("/all")
-    public ResponseEntity<List<ProfessorGetDTO>> getAllProfessors() {
-        List<Professor> professors = service.getAllProfessors();
-        return ResponseEntity.ok(professors.stream().map(ProfessorGetDTO::new).toList());
-    }
+    ResponseEntity<List<ProfessorGetDTO>> getAllProfessors();
 
     @Operation(summary = "Get a paginated list of professors")
     @ApiResponses(value = {
             @ApiResponse(
                     responseCode = "200",
                     description = "Returns a paginated list of professors",
-                    content = { @Content(mediaType = "application/json",
-                            schema = @Schema(implementation = ProfessorGetDTO.class)) })
+                    content = {@Content(mediaType = "application/json",
+                            schema = @Schema(implementation = ProfessorGetDTO.class))})
     })
-    @GetMapping
-    public ResponseEntity<Page<ProfessorGetDTO>> getPaginatedProfessors(Pageable page) {
-        Page<Professor> professors = service.getPaginatedProfessors(page);
-        return ResponseEntity.ok(professors.map(ProfessorGetDTO::new));
-    }
+    ResponseEntity<Page<ProfessorGetDTO>> getPaginatedProfessors(Pageable page);
 
     @Operation(summary = "Get a professor by its id")
     @ApiResponses(value = {
             @ApiResponse(
                     responseCode = "200",
                     description = "Returns the specified professor",
-                    content = { @Content(
+                    content = {@Content(
                             mediaType = "application/json",
-                            schema = @Schema(implementation = ProfessorGetDTO.class)) }),
+                            schema = @Schema(implementation = ProfessorGetDTO.class))}),
             @ApiResponse(
                     responseCode = "400",
                     description = "Invalid professor id",
-                    content = { @Content(
+                    content = {@Content(
                             mediaType = "application/json",
-                            schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
+                            schema = @Schema(implementation = GeneralErrorResponseDTO.class))}),
             @ApiResponse(
                     responseCode = "404",
                     description = "Professor not found",
-                    content = { @Content(
+                    content = {@Content(
                             mediaType = "application/json",
-                            schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
+                            schema = @Schema(implementation = GeneralErrorResponseDTO.class))}),
     })
-    @GetMapping("/{id}")
-    public ResponseEntity<ProfessorGetDTO> getProfessor(@PathVariable UUID id) {
-        Optional<Professor> professorById = service.getProfessorById(id);
-        return professorById.map(professor ->
-                        ResponseEntity.ok(new ProfessorGetDTO(professor)))
-                .orElseThrow(() -> new EntityNotFoundException(("Professor not found")));
-    }
+    ResponseEntity<ProfessorGetDTO> getProfessor(@PathVariable UUID id);
 
     @Operation(summary = "Create a new professor")
     @ApiResponses(value = {
             @ApiResponse(
                     responseCode = "201",
                     description = "Successfully create a professor",
-                    content = { @Content(
+                    content = {@Content(
                             mediaType = "application/json",
-                            schema = @Schema(implementation = ProfessorGetDTO.class)) }),
+                            schema = @Schema(implementation = ProfessorGetDTO.class))}),
             @ApiResponse(
                     responseCode = "400",
                     description = "Fail on request body validation",
-                    content = { @Content(
+                    content = {@Content(
                             mediaType = "application/json",
-                            schema = @Schema(implementation = ValidationErrorResponseDTO.class)) }),
+                            schema = @Schema(implementation = ValidationErrorResponseDTO.class))}),
             @ApiResponse(
                     responseCode = "409",
                     description = "Email already in use",
-                    content = { @Content(
+                    content = {@Content(
                             mediaType = "application/json",
-                            schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
+                            schema = @Schema(implementation = GeneralErrorResponseDTO.class))}),
     })
-    @PostMapping
-    public ResponseEntity<ProfessorGetDTO> saveProfessor(@Valid @RequestBody UserSignupDTO professorPostDTO) {
-        Professor newProfessor = service.save(professorPostDTO);
-        return ResponseEntity
-                .created(linkTo(methodOn(ProfessorController.class)
-                        .getProfessor(newProfessor.getId()))
-                        .toUri())
-                .body(new ProfessorGetDTO(newProfessor));
-    }
+    ResponseEntity<ProfessorGetDTO> saveProfessor(@Valid @RequestBody UserSignupDTO professorPostDTO);
 
     @Operation(summary = "Edit a professor by its id")
     @ApiResponses(value = {
             @ApiResponse(
                     responseCode = "200",
                     description = "Successfully edit the professor",
-                    content = { @Content(
+                    content = {@Content(
                             mediaType = "application/json",
-                            schema = @Schema(implementation = ProfessorGetDTO.class)) }),
+                            schema = @Schema(implementation = ProfessorGetDTO.class))}),
             @ApiResponse(
                     responseCode = "400",
                     description = "Invalid professor id",
-                    content = { @Content(
+                    content = {@Content(
                             mediaType = "application/json",
-                            schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
+                            schema = @Schema(implementation = GeneralErrorResponseDTO.class))}),
             @ApiResponse(
                     responseCode = "400",
                     description = "Fail on request body validation",
-                    content = { @Content(
+                    content = {@Content(
                             mediaType = "application/json",
-                            schema = @Schema(implementation = ValidationErrorResponseDTO.class)) }),
+                            schema = @Schema(implementation = ValidationErrorResponseDTO.class))}),
             @ApiResponse(
                     responseCode = "404",
                     description = "Professor not found",
-                    content = { @Content(
+                    content = {@Content(
                             mediaType = "application/json",
-                            schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
+                            schema = @Schema(implementation = GeneralErrorResponseDTO.class))}),
             @ApiResponse(
                     responseCode = "409",
                     description = "Email already in use",
-                    content = { @Content(
+                    content = {@Content(
                             mediaType = "application/json",
-                            schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
+                            schema = @Schema(implementation = GeneralErrorResponseDTO.class))}),
     })
-    @PutMapping("/{id}")
-    public ResponseEntity<ProfessorGetDTO> editProfessor(
+    ResponseEntity<ProfessorGetDTO> editProfessor(
             @PathVariable UUID id,
-            @Valid @RequestBody UserPutDTO userPutDTO) {
-        Professor professor = service.edit(id, userPutDTO);
-        return ResponseEntity.ok(new ProfessorGetDTO(professor));
-    }
+            @Valid @RequestBody UserPutDTO userPutDTO);
 
     @Operation(summary = "Delete a professor by its id")
     @ApiResponses(value = {
@@ -180,25 +139,21 @@ public class ProfessorController {
             @ApiResponse(
                     responseCode = "400",
                     description = "Invalid professor id",
-                    content = { @Content(
+                    content = {@Content(
                             mediaType = "application/json",
-                            schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
+                            schema = @Schema(implementation = GeneralErrorResponseDTO.class))}),
             @ApiResponse(
                     responseCode = "400",
                     description = "Cannot delete professor because they are currently assigned to one or more active classrooms",
-                    content = { @Content(
+                    content = {@Content(
                             mediaType = "application/json",
-                            schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
+                            schema = @Schema(implementation = GeneralErrorResponseDTO.class))}),
             @ApiResponse(
                     responseCode = "404",
                     description = "Professor not found",
-                    content = { @Content(
+                    content = {@Content(
                             mediaType = "application/json",
-                            schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
+                            schema = @Schema(implementation = GeneralErrorResponseDTO.class))}),
     })
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteProfessor(@PathVariable UUID id) {
-        service.deleteById(id);
-        return ResponseEntity.noContent().build();
-    }
+    ResponseEntity<Void> deleteProfessor(@PathVariable UUID id);
 }
