+++ src/main/java/com/luminabackend/controllers/task/TaskControllerDocumentation.java
@@ -1,8 +1,6 @@
-package com.luminabackend.controllers;
+package com.luminabackend.controllers.task;
 
-import com.luminabackend.models.education.task.TaskGetDTO;
 import com.luminabackend.models.education.task.*;
-import com.luminabackend.services.TaskService;
 import com.luminabackend.utils.errors.GeneralErrorResponseDTO;
 import com.luminabackend.utils.errors.ValidationErrorResponseDTO;
 import io.swagger.v3.oas.annotations.Operation;
@@ -11,18 +9,13 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.validation.Valid;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.UUID;
-
-import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
-import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 
 @ApiResponses(value = {
         @ApiResponse(
@@ -44,12 +37,7 @@ import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
                         mediaType = "application/json",
                         schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
 })
-@RestController
-@RequestMapping("/classroom/{classroomId}/task")
-public class TaskController {
-    @Autowired
-    private TaskService taskService;
-
+public interface TaskControllerDocumentation {
     @Operation(summary = "Get a list of classroom tasks based on user access")
     @ApiResponses(value = {
             @ApiResponse(
@@ -64,17 +52,9 @@ public class TaskController {
                             mediaType = "application/json",
                             schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
     })
-    @PreAuthorize("(hasRole('ADMIN') or hasRole('PROFESSOR') or hasRole('STUDENT')) and @resourceAccess.verifyClassroomAccess(#authorizationHeader, #classroomId)")
-    @GetMapping("/all")
-    public ResponseEntity<List<TaskGetDTO>> getAllClassroomTasks(
+    ResponseEntity<List<TaskGetDTO>> getAllClassroomTasks(
             @PathVariable UUID classroomId,
-            @RequestHeader("Authorization") String authorizationHeader){
-        List<TaskGetDTO> tasks = taskService.getAllTasksByClassroomId(classroomId)
-                .stream()
-                .map(TaskGetDTO::new)
-                .toList();
-        return ResponseEntity.ok(tasks);
-    }
+            @RequestHeader("Authorization") String authorizationHeader);
 
     @Operation(summary = "Get a paginated list of classroom tasks based on user access")
     @ApiResponses(value = {
@@ -90,15 +70,10 @@ public class TaskController {
                             mediaType = "application/json",
                             schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
     })
-    @PreAuthorize("(hasRole('ADMIN') or hasRole('PROFESSOR') or hasRole('STUDENT')) and @resourceAccess.verifyClassroomAccess(#authorizationHeader, #classroomId)")
-    @GetMapping
-    public ResponseEntity<Page<TaskGetDTO>> getPaginatedClassroomTasks(
+    ResponseEntity<Page<TaskGetDTO>> getPaginatedClassroomTasks(
             @PathVariable UUID classroomId,
             Pageable page,
-            @RequestHeader("Authorization") String authorizationHeader){
-        Page<Task> tasks = taskService.getPaginatedClassroomTasks(classroomId, page);
-        return ResponseEntity.ok(tasks.map(TaskGetDTO::new));
-    }
+            @RequestHeader("Authorization") String authorizationHeader);
 
     @Operation(summary = "Get a task by its id based on user access")
     @ApiResponses(value = {
@@ -120,15 +95,10 @@ public class TaskController {
                             mediaType = "application/json",
                             schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
     })
-    @PreAuthorize("(hasRole('ADMIN') or hasRole('PROFESSOR') or hasRole('STUDENT')) and @resourceAccess.verifyClassroomAccess(#authorizationHeader, #classroomId)")
-    @GetMapping("/{taskId}")
-    public ResponseEntity<TaskGetDTO> getClassroomTask(
+    ResponseEntity<TaskGetDTO> getClassroomTask(
             @PathVariable UUID classroomId,
             @PathVariable UUID taskId,
-            @RequestHeader("Authorization") String authorizationHeader){
-        Task task = taskService.getTaskById(taskId);
-        return ResponseEntity.ok(new TaskGetDTO(task));
-    }
+            @RequestHeader("Authorization") String authorizationHeader);
 
     @Operation(summary = "Create a new task")
     @ApiResponses(value = {
@@ -150,20 +120,10 @@ public class TaskController {
                             mediaType = "application/json",
                             schema = @Schema(implementation = ValidationErrorResponseDTO.class)) }),
     })
-    @PreAuthorize("(hasRole('ADMIN') or hasRole('PROFESSOR')) and @resourceAccess.verifyClassroomAccess(#authorizationHeader, #classroomId)")
-    @PostMapping
-    public ResponseEntity<TaskGetDTO> createTask(
+    ResponseEntity<TaskGetDTO> createTask(
             @PathVariable UUID classroomId,
             @Valid @RequestBody TaskPostDTO taskPostDTO,
-            @RequestHeader("Authorization") String authorizationHeader) {
-        TaskCreateDTO taskCreateDTO = new TaskCreateDTO(taskPostDTO, classroomId);
-        Task savedTask = taskService.save(taskCreateDTO);
-        return ResponseEntity
-                .created(linkTo(methodOn(TaskController.class)
-                        .getClassroomTask(classroomId, savedTask.getId(), authorizationHeader))
-                        .toUri())
-                .body(new TaskGetDTO(savedTask));
-    }
+            @RequestHeader("Authorization") String authorizationHeader);
 
     @Operation(summary = "Edit a task by its id")
     @ApiResponses(value = {
@@ -191,16 +151,11 @@ public class TaskController {
                             mediaType = "application/json",
                             schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
     })
-    @PreAuthorize("(hasRole('ADMIN') or hasRole('PROFESSOR')) and @resourceAccess.verifyClassroomAccess(#authorizationHeader, #classroomId)")
-    @PutMapping("/{taskId}")
-    public ResponseEntity<TaskGetDTO> editTask(
+    ResponseEntity<TaskGetDTO> editTask(
             @PathVariable UUID classroomId,
             @PathVariable UUID taskId,
             @Valid @RequestBody TaskPutDTO taskPutDTO,
-            @RequestHeader("Authorization") String authorizationHeader) {
-        Task task = taskService.edit(taskId, taskPutDTO);
-        return ResponseEntity.ok(new TaskGetDTO(task));
-    }
+            @RequestHeader("Authorization") String authorizationHeader);
 
     @Operation(summary = "Delete a task and its dependencies by task id")
     @ApiResponses(value = {
@@ -220,13 +175,8 @@ public class TaskController {
                             mediaType = "application/json",
                             schema = @Schema(implementation = GeneralErrorResponseDTO.class)) }),
     })
-    @PreAuthorize("(hasRole('ADMIN') or hasRole('PROFESSOR')) and @resourceAccess.verifyClassroomAccess(#authorizationHeader, #classroomId)")
-    @DeleteMapping("/{taskId}")
-    public ResponseEntity<Void> deleteTask(
+    ResponseEntity<Void> deleteTask(
             @PathVariable UUID classroomId,
             @PathVariable UUID taskId,
-            @RequestHeader("Authorization") String authorizationHeader) {
-        taskService.deleteById(taskId);
-        return ResponseEntity.noContent().build();
-    }
+            @RequestHeader("Authorization") String authorizationHeader);
 }
